backend/requirements.txt
datasets==2.20.0
Flask==3.0.3
Flask_Cors==4.0.1
numpy==2.1.0
Pillow==10.4.0
python-dotenv==1.0.1
Requests==2.32.3
tensorflow==2.16.2
